% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cacheRExample.R
\docType{package}
\name{cacheRExample}
\alias{cacheRExample}
\title{\code{cacheRExample}: An R package for post construction monitoring analyses}
\description{
Imports and stuff
}
\section{Description}{
 The \code{cacheRExample}
}

\section{\code{cacheRExample} test}{
NA
}

